+++ src/lib/components/02-organisms/CategorizedListRoot/utils.js
@@ -186,13 +186,11 @@ export const handleNodeMain = (
      * two things:
      * 1) Activate the node and its descendants.
      */
-    const a = new Date().getTime();
     changeObjects = R.uniq(
       R.flatten(
         activateNodeAndItsDescendants(node, reducedStructure, changeObjects)
       )
     );
-    console.info(`Mihin aika menee: ${(new Date().getTime() - a) / 1000} s`);
     // 2) Activate the node's predecessors.
     changeObjects = activatePredecessors(node, reducedStructure, changeObjects);
   } else if (requestedChanges.isChecked === false) {
